+++ extensions/ql-vscode/src/query-history/query-history-profile-converter.ts
